fixed linting for student app
