--- frontend/apps/student/eslint.config.js
@@ -1,43 +1,0 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import storybook from "eslint-plugin-storybook";
-import js from "@eslint/js";
-import globals from "globals";
-import reactHooks from "eslint-plugin-react-hooks";
-import reactRefresh from "eslint-plugin-react-refresh";
-import tseslint from "typescript-eslint";
-import { globalIgnores } from "eslint/config";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-export default tseslint.config(
-  [
-    globalIgnores(["dist"]),
-    {
-      files: ["**/*.{ts,tsx}"],
-      extends: [
-        js.configs.recommended,
-        tseslint.configs.recommended,
-        reactHooks.configs["recommended-latest"],
-        reactRefresh.configs.vite,
-      ],
-      languageOptions: {
-        parser: tseslint.parser,
-        parserOptions: {
-          project: [
-            "./tsconfig.json",
-            "./tsconfig.app.json",
-            "./tsconfig.node.json",
-          ],
-          // IMPORTANT: anchor the parser to this folder
-          tsconfigRootDir: __dirname,
-          ecmaVersion: 2020,
-          sourceType: "module",
-        },
-        globals: globals.browser,
-      },
-    },
-  ],
-  storybook.configs["flat/recommended"],
-  storybook.configs["flat/recommended"],
-);
